externalized run profiles, so can be run with e.g. 'SBT_OPTS="-Djava.library.path=/usr/local/lib -Xms2G -Xmx3G" sbt -Dprofile=zmq clean run'

+++ src/main/scala/org/gitpod/tick/TickRate.scala
@@ -12,6 +12,6 @@ import akka.util.Duration
 class TickRate( val current: Long,  val previous: Long, val interval: Duration ) {
 
   def ratePerSecond = {
-    ( ( current - previous ).toDouble / interval.toSeconds )
+    ( ( current - previous ) / interval.toSeconds )
   }
 }
